+++ api/src/main/kotlin/github/com/hukuta94/delivery/api/startup/configuration/adapter/orm/OrmCourierRepositoryConfiguration.kt
@@ -2,7 +2,7 @@ package github.com.hukuta94.delivery.api.startup.configuration.adapter.orm
 
 import github.com.hukuta94.delivery.infrastructure.adapter.orm.repository.jpa.CourierJpaRepository
 import github.com.hukuta94.delivery.infrastructure.adapter.orm.repository.OrmCourierRepositoryAdapter
-import github.com.hukuta94.delivery.infrastructure.adapter.orm.repository.OrmOutboxRepository
+import github.com.hukuta94.delivery.infrastructure.adapter.orm.repository.box.OrmOutboxRepository
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
